refactor: misc renamings
